Correct info_join when 'x' is a column of df x
info_join() throws an error when 'x' or 'y' is the name of a column in df x or df
y.
Solve it by changing the names of df x and y in the first lines of the
function to restore them before the full_join operation.
Improves the documentation (roxigen comments) of info_join, and
description of the package.